requirements.txt
Flask==2.2.5
PyJWT==2.8.0
mysql-connector-python==8.0.33

